\name{alpha2image}
\alias{alpha2image}
\title{Image manipulation}
\usage{
  alpha2image(img, threshold = 0.5)
}
\arguments{
  \item{img}{A rasterImage}

  \item{threshold}{An intensity threshold, all pixels
  brighter than this will be set to transparent}
}
\description{
  This function adds an alpha channel to an image
}

