% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getDatabaseIDs}
\alias{getDatabaseIDs}
\title{MetFrag result tools}
\usage{
getDatabaseIDs(mols, idprop = "Identifier")
}
\arguments{
\item{mols}{A list of rCDK \code{molecules} with MetFrag or MetFusion results}

\item{idprop}{The name of the property of the molecules where the database ID is kept}
}
\description{
Small tools to extract information from MetFrag or MetFusion results
}

