% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getScores}
\alias{getScores}
\title{MetFrag result tools}
\usage{
getScores(mols, scoreprop = "Score")
}
\arguments{
\item{mols}{A list of rCDK \code{molecules} with MetFrag or MetFusion results}

\item{scoreprop}{The name of the property of the molecules where the score is kept}
}
\description{
Small tools to extract information from MetFrag or MetFusion results
}

