% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getClusterMCSS}
\alias{getClusterMCSS}
\title{Chemical Clustering}
\usage{
getClusterMCSS(hcluster, mols, h = NULL, k = NULL, which = NULL)
}
\arguments{
\item{hcluster}{Previously created hclust for \code{mols}}

\item{mols}{The a list of rCDK \code{mols}}

\item{k,h}{Scalar. Cut the dendrogram such that either exactly
\code{k} clusters are produced or by cutting at height \code{h}.
(either k or h needs to be specified)}

\item{which}{A vector selecting the clusters to be returned}
}
\description{
Calculate the Maximum Common Substructure
}
\examples{
       smiles <- c('CCC', 'CCN', 'CCN(C)(C)',
                   'c1ccccc1Cc1ccccc1',
                   'C1CCC1CC(CN(C)(C))CC(=O)CC')
       mols <- parse.smiles(smiles)
       cluster <- hclust.mols(mols)
       clusterreps <- getClusterMCSS(cluster, mols, k=2)
}

