% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getPeaksExplained}
\alias{getPeaksExplained}
\title{MetFrag result tools}
\usage{
getPeaksExplained(mols)
}
\arguments{
\item{mols}{A list of rCDK \code{molecules} with MetFrag or MetFusion results}
}
\description{
Small tools to extract information from MetFrag or MetFusion results
}

