% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plotCluster}
\alias{plotCluster}
\title{Chemical Clustering}
\usage{
plotCluster(mols, scoreprop = "Score", idprop = "Identifier", h = NULL,
  k = NULL, ...)
}
\arguments{
\item{mols}{The a list of rCDK \code{mols}}

\item{scoreprop}{The name of the property of the molecules where the score is kept}

\item{idprop}{The name of the property of the molecules where the database ID is kept}

\item{k,h}{Scalar. Cut the dendrogram such that either exactly
\code{k} clusters are produced or by cutting at height \code{h}.
(either k or h needs to be specified)}

\item{...}{the remaining parameters are passed down to \code{dendromat()}}
}
\description{
Cluster a set of compounds and plot the results
}
\examples{
library(rcdk)
       smiles <- c('CCC', 'CCN', 'CCN(C)(C)',
                   'c1ccccc1Cc1ccccc1',
                   'C1CCC1CC(CN(C)(C))CC(=O)CC')
       mols <- parse.smiles(smiles)
       dummy <- mapply(set.property, mols, "Score", c(1,2,3,4,5))
       dummy <- mapply(set.property, mols, "Identifier", c("C1", "C2", "C3", "C4", "C5"))
       plotCluster(mols, h=0.2)
}

