% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plotMol}
\alias{plotMol}
\title{Plotting a molecule}
\usage{
plotMol(mol = NULL, smiles = NULL, width = 200, height = 200,
  watermark = NULL)
}
\arguments{
\item{mol}{An rCDK \code{mol} object}

\item{smiles}{A SMILES string, ignored if mol is supplied}

\item{width,height}{Number of pixels for molecule image in x,y}

\item{watermark}{A string written across the image as watermark}
}
\description{
rCDK provides either functions to plot into a Java window,
or into a rasterImage object. plotMol
is a high-level plotting function to plot into an R device.
}

